requirements.txt
langchain
langchain-google-genai
langchain-community
langgraph
faiss-cpu
python-dotenv
tiktoken
streamlit